Move experiment configs to own file
To enable importing them into both run models and storage without circular imports. If left in run_models module, importing from storage would introduce the circular import, as run_models import storage.
