+++ src/ert/gui/simulation/multiple_data_assimilation_panel.py
@@ -11,6 +11,7 @@ from PyQt6.QtWidgets import QCheckBox, QFormLayout, QHBoxLayout, QLabel, QWidget
 from typing_extensions import override
 
 from ert.config import ErrorInfo, ParameterConfig
+from ert.experiment_configs import MultipleDataAssimilationConfig
 from ert.gui.ertnotifier import ErtNotifier
 from ert.gui.ertwidgets import (
     ActiveRealizationsModel,
@@ -25,7 +26,7 @@ from ert.gui.ertwidgets import (
     get_parameters_button,
 )
 from ert.mode_definitions import ES_MDA_MODE
-from ert.run_models import MultipleDataAssimilation, MultipleDataAssimilationConfig
+from ert.run_models import MultipleDataAssimilation
 from ert.storage.realization_storage_state import RealizationStorageState
 from ert.validation import (
     ExperimentValidation,
