+++ src/ert/run_models/ensemble_experiment.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import logging
-from typing import ClassVar
 from uuid import UUID
 
 from pydantic import PrivateAttr
@@ -11,19 +10,14 @@ from ert.config import (
     PreExperimentFixtures,
 )
 from ert.ensemble_evaluator import EvaluatorServerConfig
+from ert.experiment_configs import EnsembleExperimentConfig
 from ert.run_models.initial_ensemble_run_model import (
     InitialEnsembleRunModel,
-    InitialEnsembleRunModelConfig,
 )
 from ert.storage import Ensemble
 from ert.trace import tracer
 
 logger = logging.getLogger(__name__)
-
-
-class EnsembleExperimentConfig(InitialEnsembleRunModelConfig):
-    target_ensemble: str
-    supports_rerunning_failed_realizations: ClassVar[bool] = True
 
 
 class EnsembleExperiment(InitialEnsembleRunModel, EnsembleExperimentConfig):
